Move property `subjects` to `Identifiable`
Ticket: mdz/webapps/internal/cudami/-/issues/545
